chore: changing the types file
