+++ src/definition/accessors/IOAuthAppsReader.ts
@@ -1,16 +1,4 @@
-export interface IOAuthApp {
-    id: string;
-    name: string;
-    active: boolean;
-    clientId?: string;
-    clientSecret?: string;
-    redirectUri: string;
-    createdAt?: string;
-    updatedAt?: string;
-    createdBy: { username: string; id: string };
-}
-
-export type IOAuthAppParams = Omit<IOAuthApp, 'id' | 'createdAt' | 'updatedAt' | 'createdBy' | 'appId'>;
+import { IOAuthApp } from './IOAuthApp';
 
 export interface IOAuthAppsReader {
     /**
